fix: Flickering Tooltip
This fixes flickering Tooltip with `trigger="hover"` when the Tooltip
overlaps the Trigger and cursor is moved to the overlaping section.
Hovering overlaping section triggers `mouseleave` on the Trigger,
which runs `hideTooltip()` but as the Tooltip disappears, the cursor
encounters the Trigger again, `mouseenter` is triggered which runs
`showTooltip()` and so it continues ad infinituum.

This solution relies on `mousemove` instead and compares the cursor
position with DOMRect of the Trigger. Also, when the `interactive` prop
is enabled, it takes DOMRect of the Tooltip into consideration.

This solution fixes another bug, which will occur with `trigger="hover"`
in conjunction with `interactive` and big offset. The offset will create
gap which will hide the Tooltip before the user hovers over it, so the
`interactive` would not work, since the Tooltip will never be hoverable.
This change virtualy increases the DOMRect of the Trigger to reach the
Tooltip DOMRect so hovering over gap keeps the Tooltip opened.